requirements.txt
Flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.18.1
Scrapy==2.11.1
newspaper3k==0.2.8
PyPDF2==3.0.1
pdfminer.six==20231228
python-whois==0.8.0
tldextract==5.1.1
duckduckgo-search==4.4.3
waybackpy==3.0.6
GitPython==3.1.42
jsonlines==4.0.0
feedparser==6.0.11
pyexcel==0.7.0
pyexcel-xlsx==0.6.0
python-dateutil==2.8.2
tabulate==0.9.0
python-docx==1.1.0 
dateparser